refactored tests: merge, roundrobin
